Connor branch2

+++ backtofront/demo/src/main/java/backtofront/example/demo/MovieRepository.java
@@ -3,7 +3,7 @@ package backtofront.example.demo;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface MovieRepository extends JpaRepository<Movie, Long> {
+public interface MovieRepository extends JpaRepository<Movie, Integer> {
 
     
     List<Movie> findByTitleContaining(String title);
